+++ include/ctl/highest.h
@@ -62,7 +62,7 @@ concept arithmetic = std::integral<T> || std::floating_point<T>;
  *   #include "ctl/highest.h"
  *   auto x0 = ctl::highest<int8_t>;  // 127
  *   auto x1 = ctl::highest<int16_t>; // 32767
- *   auto x2 = ctl::highest<float>;   // 3.40282e+38
+ *   auto x2 = ctl::highest<float>;   // 3.40283e+38
  * \endcode
  *
  * \anchor highest
@@ -79,13 +79,13 @@ inline constexpr Arithmetic highest = std::numeric_limits<Arithmetic>::max();
 
 #define CCUTL_HIGHEST_VERSION_MAJOR 0
 #define CCUTL_HIGHEST_VERSION_MINOR 3
-#define CCUTL_HIGHEST_VERSION_PATCH 0
+#define CCUTL_HIGHEST_VERSION_PATCH 1
 
 #elif CCUTL_HIGHEST_VERSION_MAJOR != 0
 #error   ccutl.highest major version mismatch
 #elif CCUTL_HIGHEST_VERSION_MINOR != 3
 #error   ccutl.highest minor version mismatch
-#elif CCUTL_HIGHEST_VERSION_PATCH != 0
+#elif CCUTL_HIGHEST_VERSION_PATCH != 1
 #warning ccutl.highest patch version mismatch
 #endif