+++ include/ctl/lowest.h
@@ -146,13 +146,13 @@ namespace CCUTL_NAMESPACE {
 
 #define CCUTL_EXISTS_CONCEPT_VERSION_MAJOR 0
 #define CCUTL_EXISTS_CONCEPT_VERSION_MINOR 3
-#define CCUTL_EXISTS_CONCEPT_VERSION_PATCH 0
+#define CCUTL_EXISTS_CONCEPT_VERSION_PATCH 1
 
 #elif CCUTL_EXISTS_CONCEPT_VERSION_MAJOR != 0
 #error   ccutl.exists_concept major version mismatch
 #elif CCUTL_EXISTS_CONCEPT_VERSION_MINOR != 3
 #error   ccutl.exists_concept minor version mismatch
-#elif CCUTL_EXISTS_CONCEPT_VERSION_PATCH != 0
+#elif CCUTL_EXISTS_CONCEPT_VERSION_PATCH != 1
 #warning ccutl.exists_concept patch version mismatch
 #endif
 
@@ -203,13 +203,13 @@ concept exists = detail::exists::impl<Template, Args...>;
 
 #define CCUTL_EXISTS_VERSION_MAJOR 0
 #define CCUTL_EXISTS_VERSION_MINOR 3
-#define CCUTL_EXISTS_VERSION_PATCH 0
+#define CCUTL_EXISTS_VERSION_PATCH 1
 
 #elif CCUTL_EXISTS_VERSION_MAJOR != 0
 #error   ccutl.exists major version mismatch
 #elif CCUTL_EXISTS_VERSION_MINOR != 3
 #error   ccutl.exists minor version mismatch
-#elif CCUTL_EXISTS_VERSION_PATCH != 0
+#elif CCUTL_EXISTS_VERSION_PATCH != 1
 #warning ccutl.exists patch version mismatch
 #endif
 
@@ -242,9 +242,9 @@ concept arithmetic = std::integral<T> || std::floating_point<T>;
  *
  * \code
  *   #include "ctl/lowest.h"
- *   auto x0 = ctl::lowest<int8_t>;  // 127
- *   auto x1 = ctl::lowest<int16_t>; // 32767
- *   auto x2 = ctl::lowest<float>;   // 3.40282e+38
+ *   auto x0 = ctl::lowest<int8_t>;  // -128
+ *   auto x1 = ctl::lowest<int16_t>; // -32768
+ *   auto x2 = ctl::lowest<float>;   // -3.40283e+38
  * \endcode
  *
  * \anchor lowest
@@ -261,13 +261,13 @@ inline constexpr Arithmetic lowest = std::numeric_limits<Arithmetic>::lowest();
 
 #define CCUTL_LOWEST_VERSION_MAJOR 0
 #define CCUTL_LOWEST_VERSION_MINOR 3
-#define CCUTL_LOWEST_VERSION_PATCH 0
+#define CCUTL_LOWEST_VERSION_PATCH 1
 
 #elif CCUTL_LOWEST_VERSION_MAJOR != 0
 #error   ccutl.lowest major version mismatch
 #elif CCUTL_LOWEST_VERSION_MINOR != 3
 #error   ccutl.lowest minor version mismatch
-#elif CCUTL_LOWEST_VERSION_PATCH != 0
+#elif CCUTL_LOWEST_VERSION_PATCH != 1
 #warning ccutl.lowest patch version mismatch
 #endif