+++ include/ctl/same.h
@@ -94,13 +94,13 @@ using targ = detail::targ::type<idx, Ts...>;
 
 #define CCUTL_TARG_VERSION_MAJOR 0
 #define CCUTL_TARG_VERSION_MINOR 3
-#define CCUTL_TARG_VERSION_PATCH 0
+#define CCUTL_TARG_VERSION_PATCH 1
 
 #elif CCUTL_TARG_VERSION_MAJOR != 0
 #error   ccutl.targ major version mismatch
 #elif CCUTL_TARG_VERSION_MINOR != 3
 #error   ccutl.targ minor version mismatch
-#elif CCUTL_TARG_VERSION_PATCH != 0
+#elif CCUTL_TARG_VERSION_PATCH != 1
 #warning ccutl.targ patch version mismatch
 #endif
 
@@ -155,13 +155,13 @@ concept same = sizeof...(Ts) > 0 and detail::same::impl<0, Ts...>();
 
 #define CCUTL_SAME_VERSION_MAJOR 0
 #define CCUTL_SAME_VERSION_MINOR 3
-#define CCUTL_SAME_VERSION_PATCH 0
+#define CCUTL_SAME_VERSION_PATCH 1
 
 #elif CCUTL_SAME_VERSION_MAJOR != 0
 #error   ccutl.same major version mismatch
 #elif CCUTL_SAME_VERSION_MINOR != 3
 #error   ccutl.same minor version mismatch
-#elif CCUTL_SAME_VERSION_PATCH != 0
+#elif CCUTL_SAME_VERSION_PATCH != 1
 #warning ccutl.same patch version mismatch
 #endif