+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 ## [Upcoming](https://github.com/jpcx/ccutl/compare/0.2.1...devel)
 
+## [0.3.1](https://github.com/jpcx/ccutl/releases/tag/0.3.1) \[[diff](https://github.com/jpcx/ccutl/compare/0.3.0...0.3.1)\]
+
+### Added
+
+- Added migration notice to README
+- Fixed docs for `ccutl.highest` and `ccutl.lowest`
+
 ## [0.3.0](https://github.com/jpcx/ccutl/releases/tag/0.3.0) \[[diff](https://github.com/jpcx/ccutl/compare/0.2.1...0.3.0)\]
 
 ### Changed